+++ tensorflow/lite/delegates/xnnpack/fully_connected_tester.cc
@@ -264,7 +264,7 @@ std::vector<char> FullyConnectedTester::CreateTfLiteModel() const {
         TensorType_INT8, /*buffer=*/1, /*name=*/0,
         CreateQuantizationParameters(
             builder, /*min=*/0, /*max=*/0,
-            builder.CreateVector<float>({filter_scale}),
+            builder.CreateVector<float>({static_cast<float>(filter_scale)}),
             builder.CreateVector<int64_t>({0}))));
   }
   tensors.emplace_back(CreateTensor(
